improvement: use ScopeGuard and remove DeferFn
