--- include/helper/scope_guard.h
@@ -1,50 +1,0 @@
-#pragma once
-
-#define __SCOPEGUARD_CONCATENATE_IMPL(s1, s2) s1##s2
-#define __SCOPEGUARD_CONCATENATE(s1, s2) __SCOPEGUARD_CONCATENATE_IMPL(s1, s2)
-
-#include <type_traits>
-
-// ScopeGuard for C++11
-namespace swoole {
-template <typename Fun>
-class ScopeGuard {
-  public:
-    ScopeGuard(Fun &&f) : _fun(std::forward<Fun>(f)), _active(true) {}
-
-    ~ScopeGuard() {
-        if (_active) {
-            _fun();
-        }
-    }
-
-    void dismiss() {
-        _active = false;
-    }
-
-    ScopeGuard() = delete;
-    ScopeGuard(const ScopeGuard &) = delete;
-    ScopeGuard &operator=(const ScopeGuard &) = delete;
-
-    ScopeGuard(ScopeGuard &&rhs) : _fun(std::move(rhs._fun)), _active(rhs._active) {
-        rhs.dismiss();
-    }
-
-  private:
-    Fun _fun;
-    bool _active;
-};
-
-namespace detail {
-enum class ScopeGuardOnExit {};
-
-template <typename Fun>
-inline ScopeGuard<Fun> operator+(ScopeGuardOnExit, Fun &&fn) {
-    return ScopeGuard<Fun>(std::forward<Fun>(fn));
-}
-}  // namespace detail
-}  // namespace swoole
-
-// Helper macro
-#define ON_SCOPE_EXIT                                                                                                  \
-    auto __SCOPEGUARD_CONCATENATE(ext_exitBlock_, __LINE__) = swoole::detail::ScopeGuardOnExit() + [&]()
